Add files via upload
The Long Calculator project is a software application developed in C++ that aims to provide advanced arithmetic operations for working with very large numbers, beyond the limitations of standard data types like int or double. This calculator is designed to handle integer arithmetic with arbitrary precision, making it suitable for a wide range of mathematical and scientific applications where precision and accuracy are paramount.